pyproject.toml
[project]
name = "lee-tools"
version = "0.1.0"
description = "Lee Editor and Hester Daemon - Development Tools for Coefficiency"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Coefficiency", email = "dev@coefficiency.net"}
]

dependencies = [
    # Core CLI
    "click>=8.0",
    "rich>=13.0",
    "httpx>=0.25",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # AI
    "mcp>=1.0.0",
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
    "claude-code-sdk>=0.0.10",
    "anthropic>=0.40.0",  # For aura test LLM calls

    # Hester Daemon
    "fastapi>=0.100",
    "uvicorn>=0.22",
    "redis>=4.5",
    "aiofiles>=23.0",
    "asyncpg>=0.29",
    "prompt_toolkit>=3.0",

    # Lee Editor
    "textual>=0.94.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "aiohttp>=3.9.0",

    # Audio processing
    "pedalboard>=0.9.0",
    "supabase>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
# For hester aura test - requires full agentic dependencies
aura-test = [
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "celery>=5.3.0",
    "instructor[google-genai]>=1.3.3",
    "elevenlabs>=1.0.0",
    "sse-starlette>=2.0.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.26.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "python-magic>=0.4.27",
    "langgraph>=0.0.20",
]

[project.scripts]
hester = "hester.cli:main"
lee = "editor.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hester", "editor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
